Use file edition instead of yaml
The group var is polluted by non yaml standard file content,
which is fine for our usage, but not for automation tooling.

Instead of fixing everything manually, the bumper will now rely
for the group vars files, to only a filestring matching.

This can be fixed later when the release version will be put in
a separate file